swap: simplify Uint64ToUint256 calls

+++ swap/protocol_test.go
@@ -245,7 +245,7 @@ func TestEmitCheque(t *testing.T) {
 	// gasPrice on testBackend == 1
 	// estimated gas costs == 50000
 	// cheque should be sent if the accumulated amount of uncashed cheques is worth more than 100000
-	balance := Uint64ToUint256(uint64(100001))
+	balance := Uint64ToUint256(100001)
 
 	if err := testDeploy(context.Background(), debitorSwap, balance.Value()); err != nil {
 		t.Fatal(err)
